apollo/requirements.txt
astropy==5.3.4
colormath==3.0.0
corner==2.2.1
Cython==3.0.7
dynesty==2.1.2
matplotlib==3.8.0
numpy==1.26.4
Pint==0.23
pint_xarray==0.3
polars==0.20.19
PyYAML==6.0.1
PyYAML==6.0.1
scipy==1.13.0
toolz==0.12.0
tqdm==4.66.2
xarray==2023.6.0
